Refactor(jjb): Split the JJB attribution parser
jjb_parser.py held 686 lines, past the 400-line reviewability limit
the aislop code-quality gate applies, with a single 586-line class
covering disk discovery, job-name expansion and project parsing.

Split it into a package at the same dotted path: models (the two
dataclasses), discovery (YAML tag registration, template and job-group
loading, JJB file lookup), expansion (turning definitions into job
names) and parser (the JJBAttribution entry point). The two halves of
the class become mixins that JJBAttribution composes, so method
resolution and every self.<method> call behave exactly as before. The
package __init__ re-exports the previous names.
